requirements.txt
flask
flask-cors
rembg
pillow
requests
gunicorn